INTERMEDIATE COMMIT.  Files are stable and work - work in progress, but this is a solid start.
- display_name not currently used in reportValue definition.  making it
a description of the pi indicator for display on the reports.  no
longer defaults to slug, but is used to describe what it means.
- as such, added descriptions of the pi indicators to the adult report
- changed display of tabular format in reports to show both percentage and
numerator and denominator
- reducted float display on graph to show no decimal places
- Changed formating of Clinic, Year, and Month
- changed report display to plot one graph for each performance indicator.
plot title now references that performance indicator, and there is a
description of the performance indicator from the display_name.  text labels
now reflect the clinic / date of the visit.  This will be used shortly
when you are able to cycle between plot types.