Updated performance indicator report calculations to fix bugs:
- Drugs dispensed appropriately wasn't being evaluated. fixed in adults, peds
pregos.
- HIV test now checks if HIV symptoms are present and the question
is left blank - if so, assigns it to poor mgmt.  For both adults and peds.
- Weight for age wasn't being assessed properly.  If weight for age or
patient weight were left blank, zscore calc was left blank.  Thats fixed.
also was looking for string when was passing in Boolean.  fixed and tested.
won't calculate any cases entered into the system before this fix is
implemented on the clinic servers.
- Code looking at a case not being closed now also checks that the outcome
selected isn't 'none'